requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
arrow==1.2.1
asgiref==3.4.1
async-timeout==4.0.1
asynctest==0.13.0
attrs==21.2.0
base58==2.1.1
beautifulsoup4==4.10.0
binaryornot==0.4.4
bitarray==1.2.2
boto3==1.20.24
botocore==1.23.24
bs4==0.0.1
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.3
cookiecutter==1.7.3
cytoolz==0.11.2
dj-database-url==0.5.0
Django==3.2.10
django-on-heroku==1.1.2
django-storages==1.12.3
django-tailwind==2.2.2
eth-abi==2.1.1
eth-account==0.5.6
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
frozenlist==1.2.0
gunicorn==20.1.0
hexbytes==0.2.2
idna==3.3
importlib-metadata==4.10.0
ipfshttpclient==0.8.0a2
Jinja2==3.0.3
jinja2-time==0.2.0
jmespath==0.10.0
jsonschema==3.2.0
lru-dict==1.1.7
MarkupSafe==2.0.1
multiaddr==0.0.9
multidict==5.2.0
netaddr==0.8.0
parsimonious==0.8.1
Pillow==8.4.0
poyo==0.5.0
protobuf==3.19.1
psycopg2-binary==2.9.2
pycryptodome==3.12.0
pyrsistent==0.18.0
python-dateutil==2.8.2
python-decouple==3.5
python-slugify==5.0.2
pytz==2021.3
requests==2.26.0
rlp==2.0.1
s3transfer==0.5.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
text-unidecode==1.3
toolz==0.11.2
typing-extensions==3.10.0.2
urllib3==1.26.7
varint==1.0.2
web3==5.25.0
websockets==9.1
whitenoise==5.3.0
yarl==1.7.2
zipp==3.6.0
